(svn r1319) Make MaxFoundFeat return the maximum feature byte found in the NFO file, not 65535.

+++ act123_classes.cpp
@@ -53,9 +53,9 @@ void act123::init(){
 
 uint act123::MaxFoundFeat()const{
 	const vector<IDarray::info>&m=defined2IDs._m;
-	uint ret=0;
+	short ret=0;
 	for(uint i=0;i<(uint)m.size();i++)
-		ret=max<uint>(ret,m[i].feature);
+		ret=max<short>(ret,m[i].feature);
 	return ret;
 }
